começo JWT Educador + controller alterada

+++ api-codando/src/main/java/school/sptech/apicodando/service/alunoService/AlunoService.java
@@ -14,8 +14,8 @@ import school.sptech.apicodando.domain.aluno.Aluno;
 import school.sptech.apicodando.mapper.AlunoMapper;
 import school.sptech.apicodando.domain.aluno.repository.AlunoRepository;
 import school.sptech.apicodando.service.alunoService.dto.AlunoCadastroDTO;
-import school.sptech.apicodando.service.autenticacao.dto.AlunoLoginDTO;
-import school.sptech.apicodando.service.autenticacao.dto.AlunoTokenDto;
+import school.sptech.apicodando.service.alunoService.dto.dtoAluno.AlunoLoginDTO;
+import school.sptech.apicodando.service.alunoService.dto.dtoAluno.AlunoTokenDto;
 
 import java.util.List;
 import java.util.Optional;
@@ -40,18 +40,25 @@ public class AlunoService {
         String senhaCriptografada = passwordEncoder.encode(novoAluno.getSenha());
         novoAluno.setSenha(senhaCriptografada);
 
-
+        if (existePorApelido(alunoCadastroDTO.getApelido())){
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Aluno já cadastrado");
+        }
 
         this.alunoRepository.save(novoAluno);
     }
 
     public void excluir(int id) {
+
+        if (!existePorId(id)){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
         alunoRepository.deleteById(id);
     }
 
-    public void atualizar(Aluno alunoAtualizado, int id){
+    public void atualizar(Aluno alunoAtualizado, int id) {
         if (!alunoRepository.existsById(id)) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"Não encontrado");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Não encontrado");
         }
         alunoAtualizado.setIdAluno(id);
         alunoRepository.save(alunoAtualizado);
@@ -62,7 +69,7 @@ public class AlunoService {
         final UsernamePasswordAuthenticationToken credentials = new UsernamePasswordAuthenticationToken(
                 usuarioLoginDto.getApelido(), usuarioLoginDto.getSenha());
 
-            final Authentication authentication = this.authenticationManager.authenticate(credentials);
+        final Authentication authentication = this.authenticationManager.authenticate(credentials);
 
         Aluno usuarioAutenticado =
                 alunoRepository.findByApelido(usuarioLoginDto.getApelido())
@@ -77,15 +84,25 @@ public class AlunoService {
         return AlunoMapper.of(usuarioAutenticado, token);
     }
 
+    //metodos de apoio
+
     public boolean existePorId(int id) {
         return alunoRepository.existsById(id);
     }
 
-    public Optional<Aluno> listarUmPorId(int id){
+    public Optional<Aluno> listarUmPorId(int id) {
         return alunoRepository.findById(id);
     }
 
-    public List<Aluno> listarTodos(){
+    public List<Aluno> listarTodos() {
         return alunoRepository.findAll();
     }
+
+    public Boolean existePorApelido(String apelido) {
+        if (alunoRepository.findByApelido(apelido).isEmpty()) {
+            return false;
+        }
+        return true;
+    }
+
 }
